split trait checkbox into button and label
