+++ gui/src/lib/app/filter-dialog/global-filter-form/global-trait-input.svelte
@@ -2,7 +2,7 @@
     import { getFilterFormContext } from '$lib/app/form-context/context'
     import type { FormControlRecord } from '$lib/app/form-context/form-control-record'
     import type { IdFilter } from '$lib/app/form-context/types'
-    import TraitCheckbox from '$lib/components/trait-checkbox.svelte'
+    import TraitCheckbox from '$lib/components/trait-checkbox/trait-checkbox.svelte'
     import { zip } from 'radash'
 
     const { form, controls } = getFilterFormContext()
