--- gui/src/lib/components/trait-checkbox.svelte
@@ -1,136 +1,0 @@
-<script lang="ts" context="module">
-    export type TraitCheckboxValue = 'included' | 'excluded' | null
-</script>
-
-<script lang="ts">
-    import * as Tooltip from '$lib/components/ui/tooltip/index.js'
-    import { TRAITS_BY_ID, TRAIT_ICONS } from '$lib/constants'
-    import CheckmarkIcon from '$lib/icons/checkmark-icon.svelte'
-    import XIcon from '$lib/icons/x-icon.svelte'
-    import TraitTooltip from './trait-tooltip.svelte'
-
-    export let id: string
-    export let value: TraitCheckboxValue
-    export let disabled = false
-    export let disabledValue: TraitCheckboxValue = 'excluded'
-    export let disabledTooltip: string = ''
-    export let portal: string | undefined = undefined
-
-    $: label = TRAITS_BY_ID[id].display_name
-    $: src = TRAIT_ICONS[id]
-    $: actualValue = disabled ? disabledValue : value
-</script>
-
-<div
-    class:disabled
-    class="root flex flex-col justify-center items-center text-center gap-[1px]"
-    class:active={actualValue !== null}
->
-    <Tooltip.Root
-        openDelay={100}
-        closeOnPointerDown={true}
-        {portal}
-        group="trait"
-        disableHoverableContent={true}
-    >
-        <Tooltip.Trigger>
-            <button
-                {disabled}
-                on:click
-                type="button"
-                class="h-12 w-12 relative select-none"
-            >
-                <!-- Hex icon -->
-                <div class="hex hover-fill p-[2px]">
-                    <div class="hex dark-fill p-[2px]">
-                        <div class="hex light-fill p-[2px]">
-                            <div class="hex dark-fill">
-                                <img class="h-[66%] w-[66%]" {src} />
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-                <!-- Selection indicator -->
-                {#if actualValue === 'included' || actualValue === 'excluded'}
-                    <div
-                        class:green={actualValue === 'included'}
-                        class:red={actualValue === 'excluded'}
-                        class="mark absolute bottom-[2px] right-[3px] rounded-full p-[2px] text-foreground"
-                    >
-                        {#if actualValue === 'included'}
-                            <CheckmarkIcon class="h-3 w-3" />
-                        {:else if actualValue === 'excluded'}
-                            <XIcon class="h-3 w-3" />
-                        {/if}
-                    </div>
-                {/if}
-            </button>
-
-            <!-- Label -->
-            <span
-                class="text-xs text-muted-foreground whitespace-nowrap"
-            >
-                {label}
-            </span>
-        </Tooltip.Trigger>
-        <Tooltip.Content class="boring-tooltip max-w-[400px]">
-            {#if disabled}
-                {disabledTooltip}
-            {:else}
-                <TraitTooltip trait_id={id} />
-            {/if}
-        </Tooltip.Content>
-    </Tooltip.Root>
-    <!-- 
-    <ConditionalTooltip disabled={!showTooltip}>
-        <span slot="tooltip">{disabledTooltip}</span>
-    </ConditionalTooltip> -->
-</div>
-
-<style lang="postcss">
-    .hex {
-        @apply h-full w-full flex justify-center items-center;
-
-        &.dark-fill {
-            background-color: #232323;
-        }
-
-        &.light-fill {
-            background-color: #6b6d6b;
-        }
-    }
-
-    .mark {
-        /* Prevent hover effect from disappearing on icon hover */
-        pointer-events: none;
-
-        &.green {
-            background-color: var(--checked-color);
-        }
-        &.red {
-            background-color: var(--unchecked-color);
-        }
-    }
-
-    /* Highlight on hover / selection */
-    button {
-        transition: all 0.2s;
-
-        &:hover .hover-fill,
-        &:focus .hover-fill {
-            filter: drop-shadow(0px 0px 20px rgba(255, 199, 46, 0.9));
-        }
-    }
-    .root:not(.active) button,
-    button[disabled] {
-        opacity: 0.3;
-
-        &:hover {
-            opacity: 0.5;
-        }
-    }
-    root:not(.disabled) .active span {
-        @apply text-foreground;
-    }
-</style>
